Move the tile loader out of the BitmapHelper
