+++ Mapsui.Rendering.Skia/BitmapHelper.cs
@@ -1,7 +1,6 @@
 using System.IO;
 using Mapsui.Extensions;
 using Mapsui.Rendering.Skia.Extensions;
-using Mapsui.Rendering.Skia.Tiling;
 using Mapsui.Utilities;
 using SkiaSharp;
 
@@ -9,25 +8,6 @@ namespace Mapsui.Rendering.Skia;
 
 public static class BitmapHelper
 {
-    public static IRenderedTile? LoadTileBitmap(byte[] data)
-    {
-        if (data.IsSKPicture())
-        {
-            return new PictureTile(SKPicture.Deserialize(data));
-        }
-
-        using var skData = SKData.CreateCopy(data);
-        var image = SKImage.FromEncodedData(skData);
-        return new ImageTile(image);
-    }
-
-    public static ImageTile? LoadBitmap(byte[] data)
-    {
-        using var skData = SKData.CreateCopy(data);
-        var image = SKImage.FromEncodedData(skData);
-        return new ImageTile(image);
-    }
-
     public static BitmapInfo? LoadBitmap(Stream stream)
     {
         if (stream.IsSvg())
